Remove pbar argument from ValidatorMixin (#10)
Remove pbar argument from ValidatorMixin

This commit introduces a `training` parameter to the constructors of various trainer classes (TLCClassificationTrainer, TLCDetectionTrainer, TLCSegmentationTrainer) to indicate the training mode. Additionally, the constructor of TLCValidatorMixin has been refactored for improved clarity and to streamline the handling of arguments, ensuring that the training state is correctly set based on the provided parameters.

---------

Co-authored-by: Frederik Mellbye <frederik.mellbye@3lc.ai>